requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
anyio==3.7.1
attrs==23.2.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
distro==1.9.0
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
gradio==3.32.0
gradio_client==0.2.5
greenlet==3.0.3
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.2
hyperframe==6.0.1
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.0.336
langcodes==3.3.0
langsmith==0.0.80
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.7.2
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.2.0
portalocker==2.8.2
preshed==3.0.9
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
qdrant-client==1.6.4
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rpds-py==0.17.1
rsa==4.9
semantic-version==2.10.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.25
srsly==2.4.8
starlette==0.35.1
tenacity==8.2.3
thinc==8.1.12
tiktoken==0.5.2
toolz==0.12.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.25.0
wasabi==1.1.2
websockets==12.0
yarl==1.9.4
